+++ onnx/backend/test/case/node/softmaxcrossentropy.py
@@ -102,7 +102,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         sce = softmaxcrossentropy(x, labels, reduction='none')
@@ -124,7 +124,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         loss, log_prob = softmaxcrossentropy(x, labels, reduction='none', get_log_prob=True)
@@ -146,7 +146,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         weights = np.array([0.9, 0.7, 0.8, 0.9, 0.9], dtype=np.float32)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -169,7 +169,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         weights = np.array([0.9, 0.7, 0.8, 0.9, 0.9], dtype=np.float32)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -192,7 +192,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         sce = softmaxcrossentropy(x, labels, reduction='sum')
@@ -214,7 +214,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         loss, log_prob = softmaxcrossentropy(x, labels, reduction='sum', get_log_prob=True)
@@ -236,7 +236,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         sce = softmaxcrossentropy(x, labels)
@@ -258,7 +258,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         loss, log_prob = softmaxcrossentropy(x, labels, get_log_prob=True)
@@ -280,7 +280,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2).astype(np.float32)
-        y = np.random.randint(0, high=5, size=(3, 2))
+        y = np.random.randint(0, high=5, size=(3, 2)).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         sce = softmaxcrossentropy(x, y)
@@ -302,7 +302,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2).astype(np.float32)
-        y = np.random.randint(0, high=5, size=(3, 2))
+        y = np.random.randint(0, high=5, size=(3, 2)).astype(np.int64)
 
         # Compute SoftmaxCrossEntropyLoss
         loss, log_prob = softmaxcrossentropy(x, y, get_log_prob=True)
@@ -324,7 +324,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         weights = np.array([0.9, 0.7, 0.8, 0.9, 0.9], dtype=np.float32)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -347,7 +347,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         weights = np.array([0.9, 0.7, 0.8, 0.9, 0.9], dtype=np.float32)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -372,7 +372,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         labels[0] = np.int64(0)
         weights = np.array([0.9, 0.7, 0.8, 0.9, 0.9], dtype=np.float32)
 
@@ -398,7 +398,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         labels[0] = np.int64(0)
         weights = np.array([0.9, 0.7, 0.8, 0.9, 0.9], dtype=np.float32)
 
@@ -424,7 +424,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         labels[0] = np.int64(2)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -449,7 +449,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, ))
+        labels = np.random.randint(0, high=5, size=(3, )).astype(np.int64)
         labels[0] = np.int64(2)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -474,7 +474,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2))
+        labels = np.random.randint(0, high=5, size=(3, 2)).astype(np.int64)
         labels[0][0] = np.int64(1)
         weights = np.array([0.2, 0.3, 0.6, 0.1, 0.5], dtype=np.float32)
 
@@ -500,7 +500,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2))
+        labels = np.random.randint(0, high=5, size=(3, 2)).astype(np.int64)
         labels[0][0] = np.int64(1)
         weights = np.array([0.2, 0.3, 0.6, 0.1, 0.5], dtype=np.float32)
 
@@ -526,7 +526,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2))
+        labels = np.random.randint(0, high=5, size=(3, 2)).astype(np.int64)
         labels[0][0] = np.int64(2)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -551,7 +551,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2))
+        labels = np.random.randint(0, high=5, size=(3, 2)).astype(np.int64)
         labels[0][0] = np.int64(2)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -576,7 +576,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2, 7).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2, 7))
+        labels = np.random.randint(0, high=5, size=(3, 2, 7)).astype(np.int64)
         labels[0][0][0] = np.int64(2)
         weights = np.array([0.2, 0.3, 0.6, 0.1, 0.5], dtype=np.float32)
 
@@ -602,7 +602,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2, 7).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2, 7))
+        labels = np.random.randint(0, high=5, size=(3, 2, 7)).astype(np.int64)
         labels[0][0][0] = np.int64(2)
         weights = np.array([0.2, 0.3, 0.6, 0.1, 0.5], dtype=np.float32)
 
@@ -628,7 +628,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2, 7).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2, 7))
+        labels = np.random.randint(0, high=5, size=(3, 2, 7)).astype(np.int64)
         labels[0][0][0] = np.int64(2)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -653,7 +653,7 @@ class SoftmaxCrossEntropyLoss(Base):
         # Define operator inputs.
         np.random.seed(0)
         x = np.random.rand(3, 5, 2, 7).astype(np.float32)
-        labels = np.random.randint(0, high=5, size=(3, 2, 7))
+        labels = np.random.randint(0, high=5, size=(3, 2, 7)).astype(np.int64)
         labels[0][0][0] = np.int64(2)
 
         # Compute SoftmaxCrossEntropyLoss
@@ -674,7 +674,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1, dim2, dim3, dim4, dim5 = 3, 5, 6, 6, 5, 3, 4
         np.random.seed(0)
         x = np.random.rand(N, C, dim1, dim2, dim3, dim4, dim5).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5))
+        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5)).astype(np.int64)
         weight = np.random.rand(C).astype(np.float32)
 
         sce = softmaxcrossentropy(x,
@@ -696,7 +696,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1, dim2, dim3, dim4, dim5 = 3, 5, 6, 6, 5, 3, 4
         np.random.seed(0)
         x = np.random.rand(N, C, dim1, dim2, dim3, dim4, dim5).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5))
+        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5)).astype(np.int64)
         weight = np.random.rand(C).astype(np.float32)
 
         loss, log_prob = softmaxcrossentropy(x,
@@ -719,7 +719,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1, dim2, dim3, dim4, dim5 = 3, 5, 6, 6, 5, 3, 4
         np.random.seed(0)
         x = np.random.rand(N, C, dim1, dim2, dim3, dim4, dim5).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5))
+        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5)).astype(np.int64)
 
         sce = softmaxcrossentropy(x,
                                 labels,
@@ -739,7 +739,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1, dim2, dim3, dim4, dim5 = 3, 5, 6, 6, 5, 3, 4
         np.random.seed(0)
         x = np.random.rand(N, C, dim1, dim2, dim3, dim4, dim5).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5))
+        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3, dim4, dim5)).astype(np.int64)
 
         loss, log_prob = softmaxcrossentropy(x,
                                 labels,
@@ -762,7 +762,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1 = 3, 5, 6
         np.random.seed(0)
         x = np.random.rand(N, C, dim1).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1))
+        labels = np.random.randint(0, high=C, size=(N, dim1)).astype(np.int64)
         labels[0][0] = -1
         weight = np.random.rand(C).astype(np.float32)
 
@@ -788,7 +788,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1 = 3, 5, 6
         np.random.seed(0)
         x = np.random.rand(N, C, dim1).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1))
+        labels = np.random.randint(0, high=C, size=(N, dim1)).astype(np.int64)
         labels[0][0] = -1
         weight = np.random.rand(C).astype(np.float32)
 
@@ -815,7 +815,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1, dim2, dim3 = 3, 5, 6, 6, 5
         np.random.seed(0)
         x = np.random.rand(N, C, dim1, dim2, dim3).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3))
+        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3)).astype(np.int64)
         labels[0][0][0][0] = -5
 
         sce = softmaxcrossentropy(x,
@@ -839,7 +839,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C, dim1, dim2, dim3 = 3, 5, 6, 6, 5
         np.random.seed(0)
         x = np.random.rand(N, C, dim1, dim2, dim3).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3))
+        labels = np.random.randint(0, high=C, size=(N, dim1, dim2, dim3)).astype(np.int64)
         labels[0][0][0][0] = -5
 
         loss, log_prob = softmaxcrossentropy(x,
@@ -864,7 +864,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C = 3, 5
         np.random.seed(0)
         x = np.random.rand(N, C).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N))
+        labels = np.random.randint(0, high=C, size=(N)).astype(np.int64)
         labels[0] = 10
         weight = np.random.rand(C).astype(np.float32)
 
@@ -890,7 +890,7 @@ class SoftmaxCrossEntropyLoss(Base):
         N, C = 3, 5
         np.random.seed(0)
         x = np.random.rand(N, C).astype(np.float32)
-        labels = np.random.randint(0, high=C, size=(N))
+        labels = np.random.randint(0, high=C, size=(N)).astype(np.int64)
         labels[0] = 10
         weight = np.random.rand(C).astype(np.float32)
